Rename bzip2 to bz2 to match the exported extension name
