+++ hphp/runtime/ext/bz2/ext_bz2.php
